+++ app/src/main/java/org/oxycblt/auxio/music/MusicViewModel.kt
@@ -80,6 +80,7 @@ class MusicViewModel @Inject constructor(private val musicRepository: MusicRepos
 
     /**
      * Create a new generic playlist.
+     *
      * @param name The name of the new playlist. If null, the user will be prompted for a name.
      */
     fun createPlaylist(name: String? = null) {